requirements.txt
# Core dependencies
streamlit==1.41.1
python-dotenv==1.0.1
pypdf==5.1.0
groq==0.11.0

# Performance monitoring
psutil==6.1.1